Tests/reorganize: move encodeQ tests to own file
