+++ test/PHPMailer/PHPMailerTest.php
@@ -862,41 +862,6 @@ EOT;
     }
 
     /**
-     * Encoding and charset tests.
-     */
-    public function testEncodings()
-    {
-        $this->Mail->CharSet = PHPMailer::CHARSET_ISO88591;
-        self::assertSame(
-            '=A1Hola!_Se=F1or!',
-            $this->Mail->encodeQ("\xa1Hola! Se\xf1or!", 'text'),
-            'Q Encoding (text) failed'
-        );
-        self::assertSame(
-            '=A1Hola!_Se=F1or!',
-            $this->Mail->encodeQ("\xa1Hola! Se\xf1or!", 'comment'),
-            'Q Encoding (comment) failed'
-        );
-        self::assertSame(
-            '=A1Hola!_Se=F1or!',
-            $this->Mail->encodeQ("\xa1Hola! Se\xf1or!", 'phrase'),
-            'Q Encoding (phrase) failed'
-        );
-        $this->Mail->CharSet = 'UTF-8';
-        self::assertSame(
-            '=C2=A1Hola!_Se=C3=B1or!',
-            $this->Mail->encodeQ("\xc2\xa1Hola! Se\xc3\xb1or!", 'text'),
-            'Q Encoding (text) failed'
-        );
-        //Strings containing '=' are a special case
-        self::assertSame(
-            'Nov=C3=A1=3D',
-            $this->Mail->encodeQ("Nov\xc3\xa1=", 'text'),
-            'Q Encoding (text) failed 2'
-        );
-    }
-
-    /**
      * Expect exceptions on bad encoding
      */
     public function testAddAttachmentEncodingException()
